pyproject.toml
[project]
name = "segment2explain-poc"
version = "0.1.0"
description = "Evidence-grounded medical imaging interpretation using MedSAM3 and MedGemma"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Segment2Explain Team" }
]
keywords = ["medical-imaging", "segmentation", "evidence-grounding", "radiology", "medsam3", "medgemma"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.24",
    "pillow>=10.0",
    "torch>=2.2",
    "transformers>=4.50.0",
    "accelerate>=0.33.0",
    "safetensors>=0.4.3",
    "sentencepiece>=0.2.0",
    "python-dotenv>=1.2.1",
    "pydicom>=3.0.1",
    "torchvision>=0.25.0",
    "xgrammar>=0.1.31",
    "fastapi>=0.128.4",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.22",
    "opencv-python-headless>=4.13.0.92",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = [
    "*.py",
    "models",
    "tools",
    "prompts",
    "utils",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=term-missing"

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    ".venv/*",
    "venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
